+++ .eslintrc.cjs
@@ -45,6 +45,7 @@ module.exports = {
         'object-shorthand': ['error', 'always'],
         'unicorn/prefer-regexp-test': 'error',
         'unicorn/no-array-for-each': 'error',
+        'unicorn/prefer-string-replace-all': 'error',
         '@typescript-eslint/prefer-for-of': 'error',
         // todo: enable
         '@typescript-eslint/no-explicit-any': 'off',
@@ -124,7 +125,12 @@ module.exports = {
         tailwindcss: {
           config: 'packages/nextra-theme-docs/tailwind.config.js',
           callees: ['cn'],
-          whitelist: ['nextra-breadcrumb', 'nextra-bleed']
+          whitelist: [
+            'nextra-breadcrumb',
+            'nextra-bleed',
+            'nextra-menu-desktop',
+            'nextra-menu-mobile'
+          ]
         }
       },
       rules: {