+++ packages/nextra/src/mdx-plugins/index.ts
@@ -7,5 +7,6 @@ export {
   remarkLinkRewrite,
   type RemarkLinkRewriteOptions
 } from './remark-link-rewrite'
-export { remarkStaticImage } from './static-image'
-export { structurize } from './structurize'
+export { remarkMdxDisableExplicitJsx } from './remark-mdx-disable-explicit-jsx'
+export { remarkStaticImage } from './remark-static-image'
+export { remarkStructurize } from './remark-structurize'